mcp_servers/google-calendar/requirements.txt
# Google Calendar MCP Server
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.116.0
